Implementation of start roll-call
